## Description The arrows for connections between dropzones are off center when created ## Expected Behavior The arrows should start at the center of each dropzone ## Current Behavior The arrows start at top left corner of the dropzone ## Steps to Reproduce Run ArrowTestArea.java 1. double click in quick success on the blue cube 2. double click in quick sucession on the red dcub 3. arrow will appear
## Failure Logs
None
## Hypothesis for Fixing the Bug
The method for getting width and height of the dropzone got overwritten in a refactoring commit, returning the improper height and width values.
Describe the test you think will verify this bug and the code fix you believe will solve this issue.

I replicate the testing envirionment that I provide, and  re-instantiate the methods that I had before to provide the proper dropzone width and height, so it's not cenetered on the top left corner.