## Description
Currently the issue is that on the Frontend, we assign "ID"s to Dropzones. These dropzones are being passed through the BoardCreator API, which means the ID's are no longer accesible.

## Expected Behavior

The ID's should be accessible, and there are no errors thrown in the creation of the Game

## Current Behavior

It currently can't create Objects, as when it tries to look up using the DropZone ID, it can't find it.

## Steps to Reproduce

Provide detailed steps for reproducing the issue.

 1. Create a Grid Object using the modal
 2. Create a Create Game Object using the frontend interface, assigning the GameObject a DropZone ID.
 3. Error will occur

## Failure Logs

java.lang.NullPointerException: Cannot invoke "Object.toString()" because the return value of "java.util.Map.get(Object)" is null

    at oogasalad.sharedDependencies.backend.ObjectFactory.constructOwnable(ObjectFactory.java:112)
    at oogasalad.sharedDependencies.backend.ObjectFactory.createOwnable(ObjectFactory.java:167)

## Hypothesis for Fixing the Bug

The issue right now is the IDManager on the backend creates ID's of it's own, since every call made to the Backend is using BoardCreator called. Using this call, we can't assign IDs. As such, different ID's are assigned to the same dropzone.