Loading Bug Fix ## Description
The File Loading doesn't output the proper response as anticipated. The issue is that the List of Componnents is outputting

## Expected Behavior

When the File is being loaded in, the Game Name is updated. As such, the files will show the contents of the current game directory.

## Current Behavior

It returns a file not found, since the File Loading only looks for a non-existant file.

## Steps to Reproduce

Provide detailed steps for reproducing the issue.

 1. Give game name to current file loader
 1. Press button
 1. Get Error

## Failure Logs

Include any relevant print/LOG statements, error messages, or stack traces.

java.io.FileNotFoundException: data\games\checker\frontend\components.json (The system cannot find the file specified)

	at java.base/java.io.FileInputStream.open0(Native Method)
	at java.base/java.io.FileInputStream.open(FileInputStream.java:216)
	at java.base/java.io.FileInputStream.<init>(FileInputStream.java:157)

## Hypothesis for Fixing the Bug

By fixing the Current File Controller to read from the two proper files, and to output do modal controller the proper objects, ie Dropzones and GameObjects.